Changed the exports on Rust.hs, and the test file Work.hs
